requirements.txt
Package             Version
------------------- -------
asgiref             3.6.0
Django              4.2.1
django-cors-headers 4.0.0
djangorestframework 3.14.0
pip                 21.2.4
psycopg2-binary     2.9.6
pytz                2023.3
setuptools          58.0.4
sqlparse            0.4.4
